new_/requirements.txt
azure-storage-blob==12.19.0
kafka-python==2.0.2
prometheus-client==0.19.0
python-dotenv==1.0.0